ROFO-199 refactor: 뒤로가기 topbar 컴포넌트화
