+++ presentation/src/main/java/com/weit2nd/presentation/ui/select/place/SelectPlaceScreen.kt
@@ -2,9 +2,7 @@ package com.weit2nd.presentation.ui.select.place
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -20,7 +18,6 @@ import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
@@ -31,23 +28,20 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.TextFieldValue
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.navigation.NavController
 import com.weit2nd.domain.model.search.Place
-import com.weit2nd.presentation.R
 import com.weit2nd.presentation.navigation.SelectPlaceMapRoutes
 import com.weit2nd.presentation.navigation.dto.PlaceDTO
 import com.weit2nd.presentation.navigation.dto.toPlace
+import com.weit2nd.presentation.ui.common.BackTopBar
 import com.weit2nd.presentation.ui.theme.Black
 import com.weit2nd.presentation.ui.theme.DarkGray
 import com.weit2nd.presentation.ui.theme.Gray2
@@ -80,7 +74,11 @@ fun SelectPlaceScreen(
 
     Scaffold(
         topBar = {
-            TopBar(modifier = Modifier.fillMaxWidth(), onClickBackBtn = vm::onClickBackBtn)
+            BackTopBar(
+                modifier = Modifier.fillMaxWidth(),
+                title = "음식점 위치 설정",
+                onClickBackBtn = vm::onClickBackBtn,
+            )
         },
         content = { innerPadding ->
             Column(
@@ -127,40 +125,6 @@ fun SelectPlaceScreen(
         key = SelectPlaceMapRoutes.SELECT_PLACE_KEY,
     ) {
         vm.onClickPlace(it.toPlace())
-    }
-}
-
-@Composable
-private fun TopBar(
-    modifier: Modifier,
-    onClickBackBtn: () -> Unit,
-) {
-    Box(
-        modifier = modifier.height(50.dp),
-        contentAlignment = Alignment.CenterStart,
-    ) {
-        Text(
-            modifier = modifier.fillMaxWidth(),
-            textAlign = TextAlign.Center,
-            text = "음식점 위치 설정",
-            style = MaterialTheme.typography.headlineSmall,
-            color = Black,
-        )
-        Row(
-            modifier = modifier,
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            IconButton(
-                modifier = Modifier.padding(start = 4.dp),
-                onClick = { onClickBackBtn() },
-            ) {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_navigate),
-                    contentDescription = "",
-                    tint = Black,
-                )
-            }
-        }
     }
 }
 
@@ -259,14 +223,6 @@ private fun SearchPlaceItem(
             style = MaterialTheme.typography.bodyMedium,
             color = Black,
         )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun TopBarPreview() {
-    RoadyFoodyTheme {
-        TopBar(modifier = Modifier) {}
     }
 }
